dataScience/ingestaPost/requirements.txt
psycopg2-binary
faker
pandas
boto3
python-dotenv